# docs/openapi.json
{
  "openapi": "3.1.1",
...
  "info": {
    "title": "VulnProof application API",
    "version": "0.1.0",
    "description": "Generated from implemented request schemas and checked against registered routes. Arc Testnet only for live funds. Fixture verification uses FIXTURE_ONLY and TRUSTED_SERVICE. Private report downloads and internal service routes are separate services. Response projections without a dedicated DTO schema remain open objects; this contract does not promise undocumented fields."
  },
  "servers": [
    {
      "url": "/",
      "description": "The configured API origin"
    }
  ],
  "paths": {
...
    "/api/v1/bounties": {
      "get": {
        "operationId": "bounties",
        "summary": "List funded fixture bounties",
        "description": "Authenticated user. All response metadata uses Cache-Control: no-store. Custom schema refinements and current role, wallet, source, and chain checks remain server requirements.",
        "security": [
          {
            "PrivyBearer": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "$schema": "https://json-schema.org/draft/2020-12/schema",
              "default": 25,
              "type": "integer",
              "minimum": 1,
              "maximum": 100
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "$schema": "https://json-schema.org/draft/2020-12/schema",
              "type": "string",
              "pattern": "^0x[0-9a-fA-F]{64}$"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The request completed. Financial state is final only where explicitly recorded as final.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "items",
                    "nextCursor"
                  ],
                  "properties": {
                    "items": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "additionalProperties": true
                      }
                    },
                    "nextCursor": {
                      "type": [
                        "string",
                        "null"
                      ]
                    }
                  },
                  "additionalProperties": true
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ApiError"
          },
          "401": {
            "$ref": "#/components/responses/ApiError"
          },
          "403": {
            "$ref": "#/components/responses/ApiError"
          },
          "404": {
            "$ref": "#/components/responses/ApiError"
          },
          "409": {
            "$ref": "#/components/responses/ApiError"
          },
          "410": {
            "$ref": "#/components/responses/ApiError"
          },
          "413": {
            "$ref": "#/components/responses/ApiError"
          },
          "415": {
            "$ref": "#/components/responses/ApiError"
          },
          "429": {
            "$ref": "#/components/responses/ApiError"
          },
          "503": {
            "$ref": "#/components/responses/ApiError"
          }
        }
      }
    },
...
    "/api/v1/bounties/{id}/uploads": {
      "post": {
        "operationId": "prepareUpload",
        "summary": "Create a claim and encrypted upload destination",
        "description": "Authenticated researcher with a currently linked claimant wallet. All response metadata uses Cache-Control: no-store. Custom schema refinements and current role, wallet, source, and chain checks remain server requirements.",
        "security": [
          {
            "PrivyBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "$schema": "https://json-schema.org/draft/2020-12/schema",
              "type": "string",
              "pattern": "^0x[0-9a-fA-F]{64}$"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "description": "Reuse this key only for identical input. The server keeps the request hash for at least 24 hours. A changed request returns 409.",
            "schema": {
              "$schema": "https://json-schema.org/draft/2020-12/schema",
              "type": "string",
              "minLength": 16,
              "maxLength": 128,
              "pattern": "^[a-zA-Z0-9:_-]+$"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "The request completed. Financial state is final only where explicitly recorded as final.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadDestination"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ApiError"
          },
          "401": {
            "$ref": "#/components/responses/ApiError"
          },
          "403": {
            "$ref": "#/components/responses/ApiError"
          },
          "404": {
            "$ref": "#/components/responses/ApiError"
          },
          "409": {
            "$ref": "#/components/responses/ApiError"
          },
          "410": {
            "$ref": "#/components/responses/ApiError"
          },
          "413": {
            "$ref": "#/components/responses/ApiError"
          },
          "415": {
            "$ref": "#/components/responses/ApiError"
          },
          "429": {
            "$ref": "#/components/responses/ApiError"
          },
          "503": {
            "$ref": "#/components/responses/ApiError"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Request_upload"
              }
            }
          }
        }
      }
    },
...
    "/api/v1/claims/me": {
      "get": {
        "operationId": "myClaims",
        "summary": "List the researcher's claims and report metadata",
        "description": "Authenticated user. All response metadata uses Cache-Control: no-store. Custom schema refinements and current role, wallet, source, and chain checks remain server requirements.",
        "security": [
          {
            "PrivyBearer": []
          }
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "The request completed. Financial state is final only where explicitly recorded as final.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "items"
                  ],
                  "properties": {
                    "items": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "additionalProperties": true
                      }
                    }
                  },
                  "additionalProperties": true
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ApiError"
          },
          "401": {
            "$ref": "#/components/responses/ApiError"
          },
          "403": {
            "$ref": "#/components/responses/ApiError"
          },
          "404": {
            "$ref": "#/components/responses/ApiError"
          },
          "409": {
            "$ref": "#/components/responses/ApiError"
          },
          "410": {
            "$ref": "#/components/responses/ApiError"
          },
          "413": {
            "$ref": "#/components/responses/ApiError"
          },
          "415": {
            "$ref": "#/components/responses/ApiError"
          },
          "429": {
            "$ref": "#/components/responses/ApiError"
          },
          "503": {
            "$ref": "#/components/responses/ApiError"
          }
        }
      }
    },
...
    "/api/v1/organizations/{id}/reports": {
      "get": {
        "operationId": "reports",
        "summary": "List organization report access states",
        "description": "Current organization OWNER or REVIEWER. All response metadata uses Cache-Control: no-store. Custom schema refinements and current role, wallet, source, and chain checks remain server requirements.",
        "security": [
          {
            "PrivyBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "$schema": "https://json-schema.org/draft/2020-12/schema",
              "type": "string",
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "$schema": "https://json-schema.org/draft/2020-12/schema",
              "default": 25,
              "type": "integer",
              "minimum": 1,
              "maximum": 100
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "$schema": "https://json-schema.org/draft/2020-12/schema",
              "type": "string",
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The request completed. Financial state is final only where explicitly recorded as final.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "items",
                    "nextCursor"
                  ],
                  "properties": {
                    "items": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "additionalProperties": true
                      }
                    },
                    "nextCursor": {
                      "type": [
                        "string",
                        "null"
                      ]
                    }
                  },
                  "additionalProperties": true
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ApiError"
          },
          "401": {
            "$ref": "#/components/responses/ApiError"
          },
          "403": {
            "$ref": "#/components/responses/ApiError"
          },
          "404": {
            "$ref": "#/components/responses/ApiError"
          },
          "409": {
            "$ref": "#/components/responses/ApiError"
          },
          "410": {
            "$ref": "#/components/responses/ApiError"
          },
          "413": {
            "$ref": "#/components/responses/ApiError"
          },
          "415": {
            "$ref": "#/components/responses/ApiError"
          },
          "429": {
            "$ref": "#/components/responses/ApiError"
          },
          "503": {
            "$ref": "#/components/responses/ApiError"
          }
        }
      }
    },
...
    "/api/v1/uploads/{id}/ciphertext": {
      "put": {
        "operationId": "uploadCiphertext",
        "summary": "Save and verify bounded ciphertext bytes",
        "description": "The upload owner before upload expiry. All response metadata uses Cache-Control: no-store. Custom schema refinements and current role, wallet, source, and chain checks remain server requirements.",
        "security": [
          {
            "PrivyBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "$schema": "https://json-schema.org/draft/2020-12/schema",
              "type": "string",
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The request completed. Financial state is final only where explicitly recorded as final.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompletedUpload"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ApiError"
          },
          "401": {
            "$ref": "#/components/responses/ApiError"
          },
          "403": {
            "$ref": "#/components/responses/ApiError"
          },
          "404": {
            "$ref": "#/components/responses/ApiError"
          },
          "409": {
            "$ref": "#/components/responses/ApiError"
          },
          "410": {
            "$ref": "#/components/responses/ApiError"
          },
          "413": {
            "$ref": "#/components/responses/ApiError"
          },
          "415": {
            "$ref": "#/components/responses/ApiError"
          },
          "429": {
            "$ref": "#/components/responses/ApiError"
          },
          "503": {
            "$ref": "#/components/responses/ApiError"
          }
        },
        "requestBody": {
          "required": true,
          "description": "Raw ciphertext bytes. The exact size and hash must match the prepared upload. No plaintext JSON or multipart body is accepted.",
          "content": {
            "application/octet-stream": {
              "schema": {
                "type": "string",
                "format": "binary",
                "maxLength": 262192
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "responses": {
      "ApiError": {
        "description": "The request failed. The error code is stable. The message contains no private provider details.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiError"
            }
          }
        }
      }
    },
    "securitySchemes": {
      "PrivyBearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "Privy access token",
        "description": "The API verifies the token and rechecks current database membership. Local test identities work only with the explicit local adapter."
      }
    },
    "schemas": {
      "ApiError": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              },
              "requestId": {
                "type": "string"
              },
              "retryable": {
                "type": "boolean"
              }
            },
            "required": [
              "code",
              "message",
              "requestId",
              "retryable"
            ],
            "additionalProperties": false
          }
        },
        "required": [
          "error"
        ],
        "additionalProperties": false
      },
...
      "UploadDestination": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "type": "object",
        "properties": {
          "uploadId": {
            "type": "string",
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
          },
          "claimId": {
            "type": "string",
            "pattern": "^0x[0-9a-fA-F]{64}$"
          },
          "uploadPath": {
            "type": "string"
          },
          "state": {
            "type": "string",
            "const": "UPLOADING"
          },
          "expiresInSeconds": {
            "type": "number",
            "const": 900
          }
        },
        "required": [
          "uploadId",
          "claimId",
          "uploadPath",
          "state",
          "expiresInSeconds"
        ]
      },
      "CompletedUpload": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "type": "object",
        "properties": {
          "uploadId": {
            "type": "string",
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
          },
          "claimId": {
            "type": "string",
            "pattern": "^0x[0-9a-fA-F]{64}$"
          },
          "state": {
            "type": "string",
            "const": "ADMISSION_PENDING"
          }
        },
        "required": [
          "uploadId",
          "claimId",
          "state"
        ]
      },
...
      "Request_upload": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "type": "object",
        "properties": {
          "keyId": {
            "type": "string",
            "pattern": "^0x[0-9a-fA-F]{64}$"
          },
          "algorithm": {
            "type": "string",
            "const": "X25519_SEALED_BOX"
          },
          "ciphertextHash": {
            "type": "string",
            "pattern": "^0x[0-9a-fA-F]{64}$"
          },
          "byteLength": {
            "type": "integer",
            "minimum": 49,
            "maximum": 262192
          },
          "claimantWalletId": {
            "type": "string",
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
          }
        },
        "required": [
          "keyId",
          "algorithm",
          "ciphertextHash",
          "byteLength",
          "claimantWalletId"
        ],
        "additionalProperties": false
      },
...
    }
  }
}